+++ src/entities/index.ts
@@ -23,3 +23,4 @@ export { PluginHook } from './plugin/plugin-hook.entity';
 export { PluginLog } from './plugin/plugin-log.entity';
 export { SystemConfig } from './config/system-config.entity';
 export { ReportTemplate } from './report/report-template.entity';
+export { RefreshToken } from './auth/refresh-token.entity';